// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsCompanys_Dtos_PacsCompanyEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Pacs_PacsCompanys_Dtos_PacsCompanyEditDto = {
    id?: string | null;
    companyName?: string | null;
    address1?: string | null;
    address2?: string | null;
    city?: string | null;
    state?: string | null;
    zip?: string | null;
    pacsCompanyTypeId?: string | null;
    description?: string | null;
};